requirements.txt
# Inwezt AI Interface - Production Dependencies
# Python 3.11+

# --- Core API & Service ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# --- AI & LLM Frameworks (Industry Standard) ---
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-google-genai>=0.0.3
langchain-mistralai>=0.1.0

# --- LLM SDKs ---
openai>=1.10.0
google-generativeai>=0.3.2
anthropic>=0.14.0
mistralai>=1.0.0

# --- Data & Analysis ---
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.36
nselib>=2.4.2
requests>=2.31.0

# --- Database & Cloud ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
boto3>=1.34.34

# --- Testing & Quality (Optional for Prod, Good for CI) ---
pytest>=7.4.0

# --- Production Hardening ---
slowapi>=0.1.9                    # Rate limiting
sentry-sdk[fastapi]>=1.40.0       # Error tracking
prometheus-fastapi-instrumentator>=6.1.0  # Metrics

# --- Scraper Specific ---
fake-useragent>=1.4.0
tenacity>=8.2.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
